-Removed unnecessary files -Merged commits -App is now in its most basic form. (Could be considered "done")
