--- app/src/main/java/android/supremeaa/todo/ActivityTodo.java
@@ -1,109 +1,0 @@
-package android.supremeaa.todo;
-
-import android.app.Activity;
-import android.app.DatePickerDialog;
-import android.app.Dialog;
-import android.app.DialogFragment;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.ArrayAdapter;
-import android.widget.Button;
-import android.widget.DatePicker;
-import android.widget.RadioButton;
-import android.widget.Spinner;
-import android.widget.TextView;
-
-import java.util.Calendar;
-import java.util.Date;
-
-
-public class ActivityTodo extends Activity {
-
-    private TextView display;
-    private Button changeDate;
-
-    private int year;
-    private int month;
-    private int day;
-
-    static final int DATE_PICKER_ID = 1111;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_todo);
-
-        display = (TextView) findViewById(R.id.display);
-        changeDate = (Button)findViewById(R.id.setDate);
-
-        final Calendar calendar = Calendar.getInstance();
-        year = calendar.get(Calendar.YEAR);
-        month = calendar.get(Calendar.MONTH);
-        day = calendar.get(Calendar.DAY_OF_MONTH);
-
-        display.setText(new StringBuilder().append(month + 1).append("-").append(day).append("-").append(year).append(" "));
-        changeDate.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                showDialog(DATE_PICKER_ID);
-            }
-        });
-
-    }
-
-    @Override
-    protected Dialog onCreateDialog(int id) {
-        switch (id){
-            case DATE_PICKER_ID:
-                return new DatePickerDialog(this, pickerListener, year, month, day);
-        }
-        return null;
-    }
-    private DatePickerDialog.OnDateSetListener pickerListener = new DatePickerDialog.OnDateSetListener(){
-        @Override
-        public void onDateSet(DatePicker view, int selectedYear, int monthOfYear, int dayOfMonth) {
-            year = selectedYear;
-            month = monthOfYear;
-            day = dayOfMonth;
-
-            display.setText(new StringBuilder().append(month + 1).append("-").append(day).append("-").append(year).append(" "));
-        }
-    };
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.todo, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-        if (id == R.id.action_settings) {
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-    public void onRadioButtonClicked(View view) {
-        // Is the button now checked?
-        boolean checked = ((RadioButton) view).isChecked();
-
-        // Check which radio button was clicked
-        switch(view.getId()) {
-            case R.id.lowP:
-                if (checked)
-
-                    break;
-            case R.id.highP:
-                if (checked)
-
-                    break;
-        }
-
-    }
-}
