--- app/src/main/res/layout/activity_todo.xml
@@ -1,40 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".TodoActivity">
-    <TextView
-        android:text="To Do List"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-    <ListView
-        android:paddingTop="20mm"
-        android:id ="@+id/listView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:layout_alignParentTop="true"
-        />
-
-    <Button
-        android:id="@+id/trash_button"
-        android:layout_width="150px"
-        android:layout_height="150px"
-        android:layout_alignParentBottom="true"
-        android:background="@drawable/trash"
-        />
-
-    <Button xmlns:android="http://schemas.android.com/apk/res/android"
-            android:id="@+id/button_send"
-            android:layout_width="100px"
-            android:layout_height="100px"
-            android:text="+"
-            android:layout_alignParentRight="true"
-            android:onClick="sendMessage" />
-
-</RelativeLayout>
