+++ app/src/main/res/layout/todo_loadout.xml
@@ -8,47 +8,38 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context=".todo">
     <Button
-
         android:id="@+id/setDate"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Set Date"
         android:onClick="showDatePicker"
-        android:layout_toLeftOf="@+id/display"
+        android:layout_toStartOf="@+id/addTask"
         android:layout_below="@+id/editText"
-        android:layout_centerHorizontal="true"
-        />
+        android:layout_toLeftOf="@+id/addTask" />
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/display"
-        android:layout_below="@+id/editText"
-        android:text=""
-        android:textSize="35dp"
-        android:layout_centerHorizontal="true"
+        android:textSize="20sp"
         android:textAppearance="?android:textAppearanceLarge"
-        />
-
+        android:layout_below="@+id/editText" />
     <EditText
         android:id="@+id/editText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:inputType="text|textCapWords"
-        android:layout_centerHorizontal="true"
-        android:hint="Task Title"
-        android:textSize="50sp"
-        android:layout_marginBottom="20dp"
-        />
-
+        android:hint="Create new task"
+        android:textSize="20sp"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true" />
     <Button
         android:layout_marginTop="30sp"
-        android:id="@+id/button1"
+        android:id="@+id/addTask"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_centerHorizontal="true"
         android:text="Add Task"
-        android:layout_below="@+id/priority"
-        />
+        android:layout_below="@+id/priority"/>
     <RadioGroup xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -57,36 +48,44 @@
         android:layout_centerHorizontal="true"
         android:layout_below="@+id/setDate"
         android:id="@+id/priority"
-        android:layout_marginTop="20dp"
-
-        >
-        <RadioButton android:id="@+id/lowP"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="1"
-            android:onClick="onRadioButtonClicked"
-            android:layout_gravity="center"
-            android:textSize="25dp"
-
-            />
-        <RadioButton android:id="@+id/highP"
-            android:textSize="25dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="2"
-            android:onClick="onRadioButtonClicked"
-            android:layout_gravity="center"
-            />
-        <RadioButton android:id="@+id/highestP"
+        android:layout_marginTop="20dp">
+        <RadioButton
+            android:id="@+id/lowP"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="3"
             android:onClick="onRadioButtonClicked"
             android:layout_gravity="center"
-            android:textSize="25dp"
-            />
-
+            android:textSize="25sp"/>
+        <RadioButton android:id="@+id/highP"
+            android:textSize="25sp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="2"
+            android:onClick="onRadioButtonClicked"
+            android:layout_gravity="center"/>
+        <RadioButton android:id="@+id/highestP"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="1"
+            android:onClick="onRadioButtonClicked"
+            android:layout_gravity="center"
+            android:textSize="25sp"/>
     </RadioGroup>
 
+    <Button
+        android:id="@+id/trash_button"
+        android:layout_width="75dp"
+        android:layout_height="75dp"
+        android:layout_alignParentBottom="true"
+        android:background="@drawable/trash"/>
+
+    <ListView
+        android:paddingTop="20dp"
+        android:id ="@+id/listView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true"
+        android:layout_below="@+id/addTask" />
 
 </RelativeLayout>
